@@ -20,16 +20,16 @@
   ],
   "devDependencies": {
     "@esri/application-base-js": "^0.0.2",
-    "@types/arcgis-js-api": "^4.12.0",
+    "@types/arcgis-js-api": "^4.14.0",
     "autoprefixer": "^6.3.6",
     "calcite-web": "github:Esri/calcite-web#v1.2.5",
     "cpx": "^1.5.0",
     "dojo-typings": "^1.11.9",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.13.1",
     "npm-run-all": "^4.1.3",
     "postcss-normalize-charset": "^1.1.0",
-    "tslint": "^5.11.0",
-    "typescript": "^2.9.2",
+    "tslint": "^5.20.1",
+    "typescript": "^3.7.5",
     "workbox-build": "^4.3.1"
   },
   "scripts": {
@@ -41,7 +41,7 @@
     "deploy": "npm run build:sass & npm run build:ts"
   },
   "dependencies": {
-    "@esri/configurable-app-components": "^1.0.7",
-    "@esri/telemetry": "^1.4.0"
+    "@esri/configurable-app-components": "^1.1.0",
+    "@esri/telemetry": "^1.5.1"
   }
 }
