@@ -1,51 +1,36 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 /// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
 /// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
 
 // esri.core
-import Accessor = require('esri/core/Accessor');
-import Collection = require('esri/core/Collection');
-import requireUtils = require('esri/core/requireUtils');
-import promiseUtils = require('esri/core/promiseUtils');
+import Accessor = require("esri/core/Accessor");
+import Collection = require("esri/core/Collection");
 
 // esri.core.accessorSupport
-import { subclass, declared, property } from 'esri/core/accessorSupport/decorators';
+import {
+	subclass,
+	declared,
+	property
+} from "esri/core/accessorSupport/decorators";
 
 // esri.views
-import MapView = require('esri/views/MapView');
-import SceneView = require('esri/views/SceneView');
+import MapView = require("esri/views/MapView");
+import SceneView = require("esri/views/SceneView");
 
 //esri.geometry
-import Point = require('esri/geometry/Point');
+import Point = require("esri/geometry/Point");
 
 // esri.request
-import esriRequest = require('esri/request');
-
-// require
-import moduleRequire = require('require');
+import esriRequest = require("esri/request");
 
 // Share Item
-import ShareItem = require('./ShareItem');
+import ShareItem = require("./ShareItem");
 
 // Share Feature
-import ShareFeatures = require('./ShareFeatures');
+import ShareFeatures = require("./ShareFeatures");
 
-// esri.core.watchUtils
-import watchUtils = require('esri/core/watchUtils');
-
-// esri.core.Handles
-import Handles = require('esri/core/Handles');
+// esri.geometry
+import projection = require("esri/geometry/projection");
+import SpatialReference = require("esri/geometry/SpatialReference");
 
 //----------------------------------
 //
@@ -60,14 +45,25 @@ const ShareItemCollection = Collection.ofType<ShareItem>(ShareItem);
 //
 //----------------------------------
 const FACEBOOK_ITEM = new ShareItem({
-	id: 'facebook',
-	name: 'Facebook',
-	urlTemplate: 'https://www.facebook.com/sharer/sharer.php?s=100&p[url]={url}'
+	id: "facebook",
+	name: "Facebook",
+	urlTemplate: "https://www.facebook.com/sharer/sharer.php?s=100&u={url}"
 });
 const TWITTER_ITEM = new ShareItem({
-	id: 'twitter',
-	name: 'Twitter',
-	urlTemplate: 'https://twitter.com/intent/tweet?text={summary}&url={url}'
+	id: "twitter",
+	name: "Twitter",
+	urlTemplate: "https://twitter.com/intent/tweet?text={summary}&url={url}"
+});
+const LINKEDIN_ITEM = new ShareItem({
+	id: "linkedin",
+	name: "LinkedIn",
+	urlTemplate:
+		"https://www.linkedin.com/shareArticle?url={url}&title={title}&summary={summary}"
+});
+const EMAIL_ITEM = new ShareItem({
+	id: "email",
+	name: "E-mail",
+	urlTemplate: "mailto:?subject={title}&body={summary}%20{url}"
 });
 
 //----------------------------------
@@ -75,19 +71,17 @@ const TWITTER_ITEM = new ShareItem({
 //  Shorten URL API
 //
 //----------------------------------
-const SHORTEN_API = 'https://arcg.is/prod/shorten';
+const SHORTEN_API = "https://arcg.is/prod/shorten";
 
 //----------------------------------
 //
 //  State
 //
 //----------------------------------
-type State = 'ready' | 'loading' | 'shortening' | 'projecting' | 'disabled';
+type State = "ready" | "loading" | "shortening" | "projecting" | "disabled";
 
-@subclass('ShareViewModel')
+@subclass("ShareViewModel")
 class ShareViewModel extends declared(Accessor) {
-	private _handles: Handles = new Handles();
-
 	//----------------------------------
 	//
 	//  Lifecycle
@@ -105,9 +99,9 @@ class ShareViewModel extends declared(Accessor) {
 	//
 	//----------------------------------
 
-	// Promises for widget state
-	private _shortenPromise: IPromise = null;
-	private _projectionPromise: IPromise = null;
+	// To keep track of widget state
+	private _shortening = false;
+	private _projecting = false;
 
 	//----------------------------------
 	//
@@ -115,14 +109,20 @@ class ShareViewModel extends declared(Accessor) {
 	//
 	//----------------------------------
 	@property({
-		dependsOn: [ 'view.ready' ],
+		dependsOn: ["view.ready"],
 		readOnly: true
 	})
 	get state(): State {
-		const ready = this.get('view.ready');
+		const ready = this.get("view.ready");
 		return ready
-			? this._projectionPromise ? 'projecting' : this._shortenPromise ? 'shortening' : 'ready'
-			: this.view ? 'loading' : 'disabled';
+			? this._projecting
+				? "projecting"
+				: this._shortening
+					? "shortening"
+					: "ready"
+			: this.view
+				? "loading"
+				: "disabled";
 	}
 
 	//----------------------------------
@@ -131,12 +131,11 @@ class ShareViewModel extends declared(Accessor) {
 	//
 	//----------------------------------
 	@property({
-		dependsOn: [ 'shareUrl' ],
+		dependsOn: ["shareUrl"],
 		readOnly: true
 	})
 	get embedCode(): string {
-		return `<iframe src="${this
-			.shareUrl}" width="600" height="450" frameborder="0" style="border:0" allowfullscreen></iframe>`;
+		return `<iframe src="${this.shareUrl}" width="600" height="450" frameborder="0" style="border:0" allowfullscreen></iframe>`;
 	}
 
 	//----------------------------------
@@ -161,34 +160,18 @@ class ShareViewModel extends declared(Accessor) {
 
 	//----------------------------------
 	//
-	// defaultObjectId
-	//
-	//----------------------------------
-	@property() defaultObjectId: number = null;
-
-	//----------------------------------
-	//
-	// attachmentIndex
-	//
-	//----------------------------------
-	@property() attachmentIndex: number = null;
-
-	//----------------------------------
-	//
-	// isDefault
-	//
-	//----------------------------------
-	@property() isDefault: boolean = null;
-
-	//----------------------------------
-	//
 	//  shareItems
 	//
 	//----------------------------------
 	@property({
 		type: ShareItemCollection
 	})
-	shareItems: Collection<ShareItem> = new ShareItemCollection([ FACEBOOK_ITEM, TWITTER_ITEM ]);
+	shareItems: Collection<ShareItem> = new ShareItemCollection([
+		FACEBOOK_ITEM,
+		TWITTER_ITEM,
+		LINKEDIN_ITEM,
+		EMAIL_ITEM
+	]);
 
 	//----------------------------------
 	//
@@ -208,48 +191,21 @@ class ShareViewModel extends declared(Accessor) {
 	@property({ readOnly: true })
 	shareUrl: string = null;
 
-	initialize() {
-		this._handles.add([
-			watchUtils.watch(this, [ 'defaultObjectId', 'attachmentIndex' ], () => {
-				if (!this.isDefault) {
-					this.generateUrl();
-				}
-			})
-		]);
-	}
-
 	//----------------------------------
 	//
 	//  Public Methods
 	//
 	//----------------------------------
-	generateUrl(): IPromise<string> {
-		if (!this.isDefault) {
-			return this._generateShareUrl().then((url) => {
-				const { shortenLink } = this.shareFeatures;
-				if (shortenLink) {
-					return this._shorten(url).then((shortenedUrl) => {
-						this._shortenPromise = null;
-						this.notifyChange('state');
-						this._set('shareUrl', shortenedUrl);
-						return promiseUtils.resolve(shortenedUrl);
-					});
-				}
-				this._set('shareUrl', url);
-				return promiseUtils.resolve(url);
-			});
-		} else {
-			const { shortenLink } = this.shareFeatures;
-			if (shortenLink) {
-				const { href } = window.location;
-				return this._shorten(href).then((shortenedUrl) => {
-					this._shortenPromise = null;
-					this.notifyChange('state');
-					this._set('shareUrl', shortenedUrl);
-					return promiseUtils.resolve(shortenedUrl);
-				});
-			}
+	async generateUrl(): Promise<string> {
+		const url = await this._generateShareUrl();
+		const { shortenLink } = this.shareFeatures;
+		if (shortenLink) {
+			const shortenedUrl = await this._shorten(url);
+			this._set("shareUrl", shortenedUrl);
+			return shortenedUrl;
 		}
+		this._set("shareUrl", url);
+		return url;
 	}
 
 	//----------------------------------
@@ -257,11 +213,11 @@ class ShareViewModel extends declared(Accessor) {
 	//  Private Methods
 	//
 	//----------------------------------
-	private _generateShareUrl(): IPromise<string> {
+	private async _generateShareUrl(): Promise<string> {
 		const { href } = window.location;
 		// If view is not ready
-		if (!this.get('view.ready')) {
-			return promiseUtils.resolve(href);
+		if (!this.get("view.ready")) {
+			return href;
 		}
 		// Use x/y values and the spatial reference of the view to instantiate a geometry point
 		const { x, y } = this.view.center;
@@ -272,41 +228,37 @@ class ShareViewModel extends declared(Accessor) {
 			spatialReference
 		});
 		// Use pointToConvert to project point. Once projected, pass point to generate the share URL parameters
-		return this._processPoint(centerPoint).then((point: Point) => {
-			this._projectionPromise = null;
-			this.notifyChange('state');
-			return this._generateShareUrlParams(point);
-		});
+		const point = await this._processPoint(centerPoint);
+		return this._generateShareUrlParams(point);
 	}
 
-	private _processPoint(point: Point): IPromise<Point> {
+	private async _processPoint(point: Point): Promise<__esri.Point> {
 		const { isWGS84, isWebMercator } = point.spatialReference;
 		// If spatial reference is WGS84 or Web Mercator, use longitude/latitude values to generate the share URL parameters
 		if (isWGS84 || isWebMercator) {
-			return promiseUtils.resolve(point);
+			return point;
 		}
-		this._projectionPromise = requireUtils.when(moduleRequire, [
-			'esri/geometry/projection',
-			'esri/geometry/SpatialReference'
-		]);
-		this.notifyChange('state');
-		return this._projectionPromise.then(([ projection, SpatialReference ]) => {
-			const outputSpatialReference = new SpatialReference({
-				wkid: 4326
+		// Check if client side projection is not supported
+		if (!projection.isSupported()) {
+			const point = new Point({
+				x: null,
+				y: null
 			});
-			return projection.load().then(() => {
-				// Check if client side projection is not supported
-				if (!projection.isSupported()) {
-					const point = new Point({
-						x: null,
-						y: null
-					});
-					return promiseUtils.resolve(point);
-				}
-				const projectedPoint = projection.project(point, outputSpatialReference);
-				return promiseUtils.resolve(projectedPoint);
-			});
+			return point;
+		}
+		const outputSpatialReference = new SpatialReference({
+			wkid: 4326
 		});
+		this._projecting = true;
+		this.notifyChange("state");
+		await projection.load();
+		const projectedPoint = projection.project(
+			point,
+			outputSpatialReference
+		) as __esri.Point;
+		this._projecting = false;
+		this.notifyChange("state");
+		return projectedPoint;
 	}
 
 	private _generateShareUrlParams(point: Point): string {
@@ -319,16 +271,18 @@ class ShareViewModel extends declared(Accessor) {
 		const roundedLat = this._roundValue(latitude);
 		const { zoom } = this.view;
 		const roundedZoom = this._roundValue(zoom);
-		const path = href.split('center')[0];
+		const path = href.split("center")[0];
 		// If no "?", then append "?". Otherwise, check for "?" and "="
-		const sep = path.indexOf('?') === -1 ? '?' : path.indexOf('?') !== -1 && path.indexOf('=') !== -1 ? '&' : '';
-		const defaultObjectId = this.defaultObjectId !== null ? `&defaultObjectId=${this.defaultObjectId}` : '';
-
-		const attachmentIndex = this.attachmentIndex !== null ? `&attachmentIndex=${this.attachmentIndex}` : '';
-		const shareParams = `${path}${sep}center=${roundedLon},${roundedLat}&level=${roundedZoom}${defaultObjectId}${attachmentIndex}`;
-		const type = this.get('view.type');
+		const sep =
+			path.indexOf("?") === -1
+				? "?"
+				: path.indexOf("?") !== -1 && path.indexOf("=") !== -1
+					? "&"
+					: "";
+		const shareParams = `${path}${sep}center=${roundedLon},${roundedLat}&level=${roundedZoom}`;
+		const type = this.get("view.type");
 		// Checks if view.type is 3D, if so add, 3D url parameters
-		if (type === '3d') {
+		if (type === "3d") {
 			const { camera } = this.view as SceneView;
 			const { heading, fov, tilt } = camera;
 			const roundedHeading = this._roundValue(heading);
@@ -336,30 +290,26 @@ class ShareViewModel extends declared(Accessor) {
 			const roundedTilt = this._roundValue(tilt);
 			return `${shareParams}&heading=${roundedHeading}&fov=${roundedFov}&tilt=${roundedTilt}`;
 		}
+
 		// Otherwise, just return original url parameters for 2D
 		return shareParams;
 	}
 
-	private _shorten(url: string): IPromise<string> {
-		const requestOptions = {
-			callbackParamName: 'callback',
+	private async _shorten(url: string): Promise<string> {
+		this._shortening = true;
+		this.notifyChange("state");
+		const request = await esriRequest(SHORTEN_API, {
 			query: {
 				longUrl: url,
-				f: 'json'
+				f: "json"
 			}
-		} as any;
-		this._shortenPromise = esriRequest(SHORTEN_API, requestOptions);
-		this.notifyChange('state');
-		return this._shortenPromise
-			.catch((res) => {
-				return res;
-			})
-			.then((res) => {
-				const shortUrl = res.data && res.data.data && res.data.data.url;
-				if (shortUrl) {
-					return shortUrl;
-				}
-			});
+		});
+		this._shortening = false;
+		this.notifyChange("state");
+		const shortUrl = request.data && request.data.data && request.data.data.url;
+		if (shortUrl) {
+			return shortUrl;
+		}
 	}
 
 	private _roundValue(val: number): number {
