@@ -1,24 +1,3 @@
-/*
-  Copyright 2017 Esri
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-
-  you may not use this file except in compliance with the License.
-
-  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-
-  distributed under the License is distributed on an "AS IS" BASIS,
-
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-
-  See the License for the specific language governing permissions and
-
-  limitations under the License.​
-*/
 var __assign = (this && this.__assign) || Object.assign || function(t) {
     for (var s, i = 1, n = arguments.length; i < n; i++) {
         s = arguments[i];
@@ -27,9 +6,56 @@ var __assign = (this && this.__assign) || Object.assign || function(t) {
     }
     return t;
 };
-define(["require", "exports", "esri/core/requireUtils", "esri/core/promiseUtils", "esri/core/watchUtils", "./urlUtils"], function (require, exports, requireUtils, promiseUtils, watchUtils, urlUtils_1) {
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (Object.hasOwnProperty.call(mod, k)) result[k] = mod[k];
+    result["default"] = mod;
+    return result;
+};
+define(["require", "exports", "esri/core/promiseUtils", "esri/core/watchUtils", "esri/views/MapView", "esri/views/SceneView", "./urlUtils"], function (require, exports, promiseUtils_1, watchUtils_1, MapView_1, SceneView_1, urlUtils_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
+    MapView_1 = __importDefault(MapView_1);
+    SceneView_1 = __importDefault(SceneView_1);
     //--------------------------------------------------------------------------
     //
     //  Public Methods
@@ -37,7 +63,9 @@ define(["require", "exports", "esri/core/requireUtils", "esri/core/promiseUtils"
     //--------------------------------------------------------------------------
     function getConfigViewProperties(config) {
         var center = config.center, components = config.components, extent = config.extent, level = config.level, viewpoint = config.viewpoint;
-        var ui = components ? { ui: { components: urlUtils_1.parseViewComponents(components) } } : null;
+        var ui = components
+            ? { ui: { components: urlUtils_1.parseViewComponents(components) } }
+            : null;
         var cameraProps = viewpoint ? { camera: urlUtils_1.parseViewpoint(viewpoint) } : null;
         var centerProps = center ? { center: urlUtils_1.parseCenter(center) } : null;
         var zoomProps = level ? { zoom: urlUtils_1.parseLevel(level) } : null;
@@ -46,51 +74,81 @@ define(["require", "exports", "esri/core/requireUtils", "esri/core/promiseUtils"
     }
     exports.getConfigViewProperties = getConfigViewProperties;
     function createView(properties) {
-        var map = properties.map;
-        if (!map) {
-            return promiseUtils.reject("properties does not contain a \"map\"");
-        }
-        var isWebMap = map.declaredClass === "esri.WebMap";
-        var isWebScene = map.declaredClass === "esri.WebScene";
-        if (!isWebMap && !isWebScene) {
-            return promiseUtils.reject("map is not a \"WebMap\" or \"WebScene\"");
-        }
-        var viewTypePath = isWebMap ? "esri/views/MapView" : "esri/views/SceneView";
-        return requireUtils.when(require, viewTypePath).then(function (ViewType) {
-            return new ViewType(properties);
+        return __awaiter(this, void 0, void 0, function () {
+            var map, isWebMap, isWebScene;
+            return __generator(this, function (_a) {
+                map = properties.map;
+                if (!map) {
+                    return [2 /*return*/, promiseUtils_1.reject("properties does not contain a \"map\"")];
+                }
+                isWebMap = map.declaredClass === "esri.WebMap";
+                isWebScene = map.declaredClass === "esri.WebScene";
+                if (!isWebMap && !isWebScene) {
+                    return [2 /*return*/, promiseUtils_1.reject("map is not a \"WebMap\" or \"WebScene\"")];
+                }
+                return [2 /*return*/, isWebMap ? new MapView_1.default(properties) : new SceneView_1.default(properties)];
+            });
         });
     }
     exports.createView = createView;
     function createMapFromItem(options) {
-        var item = options.item, appProxies = options.appProxies;
+        var item = options.item;
         var isWebMap = item.type === "Web Map";
         var isWebScene = item.type === "Web Scene";
         if (!isWebMap && !isWebScene) {
-            return promiseUtils.reject();
+            return promiseUtils_1.reject();
         }
-        return isWebMap ? createWebMapFromItem(options) : createWebSceneFromItem(options);
+        return isWebMap
+            ? createWebMapFromItem(options)
+            : createWebSceneFromItem(options);
     }
     exports.createMapFromItem = createMapFromItem;
     function createWebMapFromItem(options) {
-        var item = options.item, appProxies = options.appProxies;
-        return requireUtils.when(require, "esri/WebMap").then(function (WebMap) {
-            var wm = new WebMap({
-                portalItem: item
-            });
-            return wm.load().then(function () {
-                return _updateProxiedLayers(wm, appProxies);
+        return __awaiter(this, void 0, void 0, function () {
+            var item, appProxies, WebMap, wm;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        item = options.item, appProxies = options.appProxies;
+                        return [4 /*yield*/, new Promise(function (resolve_1, reject_1) { require(["esri/WebMap"], resolve_1, reject_1); }).then(__importStar)];
+                    case 1:
+                        WebMap = _a.sent();
+                        wm = new WebMap.default({
+                            portalItem: item
+                        });
+                        return [4 /*yield*/, wm.load()];
+                    case 2:
+                        _a.sent();
+                        return [4 /*yield*/, wm.basemap.load()];
+                    case 3:
+                        _a.sent();
+                        return [2 /*return*/, _updateProxiedLayers(wm, appProxies)];
+                }
             });
         });
     }
     exports.createWebMapFromItem = createWebMapFromItem;
     function createWebSceneFromItem(options) {
-        var item = options.item, appProxies = options.appProxies;
-        return requireUtils.when(require, "esri/WebScene").then(function (WebScene) {
-            var ws = new WebScene({
-                portalItem: item
-            });
-            return ws.load().then(function () {
-                return _updateProxiedLayers(ws, appProxies);
+        return __awaiter(this, void 0, void 0, function () {
+            var item, appProxies, WebScene, ws;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        item = options.item, appProxies = options.appProxies;
+                        return [4 /*yield*/, new Promise(function (resolve_2, reject_2) { require(["esri/WebScene"], resolve_2, reject_2); }).then(__importStar)];
+                    case 1:
+                        WebScene = _a.sent();
+                        ws = new WebScene.default({
+                            portalItem: item
+                        });
+                        return [4 /*yield*/, ws.load()];
+                    case 2:
+                        _a.sent();
+                        return [4 /*yield*/, ws.basemap.load()];
+                    case 3:
+                        _a.sent();
+                        return [2 /*return*/, _updateProxiedLayers(ws, appProxies)];
+                }
             });
         });
     }
@@ -102,28 +160,52 @@ define(["require", "exports", "esri/core/requireUtils", "esri/core/promiseUtils"
     }
     exports.getItemTitle = getItemTitle;
     function goToMarker(marker, view) {
-        if (!marker || !view) {
-            return promiseUtils.resolve();
-        }
-        return urlUtils_1.parseMarker(marker).then(function (graphic) {
-            view.graphics.add(graphic);
-            var view2 = view; // todo: Typings will be fixed in next release.
-            return view2.goTo(graphic);
+        return __awaiter(this, void 0, void 0, function () {
+            var graphic;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        if (!marker || !view) {
+                            return [2 /*return*/, promiseUtils_1.resolve()];
+                        }
+                        return [4 /*yield*/, urlUtils_1.parseMarker(marker)];
+                    case 1:
+                        graphic = _a.sent();
+                        return [4 /*yield*/, view.when()];
+                    case 2:
+                        _a.sent();
+                        view.graphics.add(graphic);
+                        view.goTo(graphic);
+                        return [2 /*return*/, graphic];
+                }
+            });
         });
     }
     exports.goToMarker = goToMarker;
     function findQuery(query, view) {
-        // ?find=redlands, ca
-        if (!query || !view) {
-            return promiseUtils.resolve();
-        }
-        return requireUtils.when(require, "esri/widgets/Search/SearchViewModel").then(function (SearchViewModel) {
-            var searchVM = new SearchViewModel({
-                view: view
-            });
-            return searchVM.search(query).then(function (result) {
-                watchUtils.whenFalseOnce(view, "popup.visible", function () { return searchVM.destroy(); });
-                return result;
+        return __awaiter(this, void 0, void 0, function () {
+            var SearchViewModel, searchVM, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        // ?find=redlands, ca
+                        if (!query || !view) {
+                            return [2 /*return*/, promiseUtils_1.resolve()];
+                        }
+                        return [4 /*yield*/, new Promise(function (resolve_3, reject_3) { require(["esri/widgets/Search/SearchViewModel"], resolve_3, reject_3); }).then(__importStar)];
+                    case 1:
+                        SearchViewModel = _a.sent();
+                        searchVM = new SearchViewModel.default({
+                            view: view
+                        });
+                        return [4 /*yield*/, searchVM.search(query)];
+                    case 2:
+                        result = _a.sent();
+                        watchUtils_1.whenFalseOnce(view, "popup.visible", function () {
+                            searchVM.destroy();
+                        });
+                        return [2 /*return*/, result];
+                }
             });
         });
     }
@@ -138,7 +220,7 @@ define(["require", "exports", "esri/core/requireUtils", "esri/core/promiseUtils"
             return webItem;
         }
         appProxies.forEach(function (proxy) {
-            webItem.layers.forEach(function (layer) {
+            webItem.allLayers.forEach(function (layer) {
                 if (layer.url === proxy.sourceUrl) {
                     layer.url = proxy.proxyUrl;
                 }
