@@ -27,9 +27,57 @@ var __assign = (this && this.__assign) || Object.assign || function(t) {
     }
     return t;
 };
-define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/core/requireUtils", "esri/geometry/Extent", "esri/geometry/Point"], function (require, exports, Camera, promiseUtils, requireUtils, Extent, Point) {
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (Object.hasOwnProperty.call(mod, k)) result[k] = mod[k];
+    result["default"] = mod;
+    return result;
+};
+define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/geometry/Extent", "esri/geometry/Point"], function (require, exports, Camera_1, promiseUtils_1, Extent_1, Point_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
+    Camera_1 = __importDefault(Camera_1);
+    Extent_1 = __importDefault(Extent_1);
+    Point_1 = __importDefault(Point_1);
     //--------------------------------------------------------------------------
     //
     //  Public Methods
@@ -54,7 +102,7 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
         var tiltAndHeadingString = viewpointArray[tiltAndHeadingIndex];
         var cameraProperties = _getCameraProperties(cameraString, tiltAndHeadingString);
         if (cameraProperties.position) {
-            return new Camera(cameraProperties);
+            return new Camera_1.default(cameraProperties);
         }
         return;
     }
@@ -78,7 +126,7 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
             return null;
         }
         var wkid = centerLength === 3 ? parseInt(centerArray[2], 10) : 4326;
-        return new Point({
+        return new Point_1.default({
             x: x,
             y: y,
             spatialReference: {
@@ -109,7 +157,7 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
             return null;
         }
         var wkid = extentLength === 5 ? parseInt(extentArray[4], 10) : 4326;
-        var ext = new Extent({
+        var ext = new Extent_1.default({
             xmin: xmin,
             ymin: ymin,
             xmax: xmax,
@@ -122,64 +170,75 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
     }
     exports.parseExtent = parseExtent;
     function parseMarker(marker) {
-        // ?marker=-117;34;4326;My Title;http://www.daisysacres.com/images/daisy_icon.gif;My location&level=10
-        // ?marker=-117,34,4326,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
-        // ?marker=-13044705.25,4036227.41,102100,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
-        // ?marker=-117,34,,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
-        // ?marker=-117,34,,,,My location&level=10
-        // ?marker=-117,34&level=10
-        // ?marker=10406557.402,6590748.134,2526
-        if (!marker) {
-            return promiseUtils.resolve();
-        }
-        var markerArray = _splitURLString(marker);
-        var markerLength = markerArray.length;
-        if (markerLength < 2) {
-            return promiseUtils.reject();
-        }
-        return requireUtils.when(require, [
-            "esri/Graphic",
-            "esri/PopupTemplate",
-            "esri/symbols/PictureMarkerSymbol",
-            "esri/symbols/SimpleMarkerSymbol"
-        ]).then(function (modules) {
-            var Graphic = modules[0], PopupTemplate = modules[1], PictureMarkerSymbol = modules[2], SimpleMarkerSymbol = modules[3];
-            var x = parseFloat(markerArray[0]);
-            var y = parseFloat(markerArray[1]);
-            var content = markerArray[3];
-            var icon_url = markerArray[4];
-            var label = markerArray[5];
-            var wkid = markerArray[2] ? parseInt(markerArray[2], 10) : 4326;
-            var markerSymbol = icon_url ? new PictureMarkerSymbol({
-                url: icon_url,
-                height: "32px",
-                width: "32px"
-            }) : new SimpleMarkerSymbol({
-                outline: {
-                    width: 1
-                },
-                size: 14,
-                color: [255, 255, 255, 0]
-            });
-            var point = new Point({
-                "x": x,
-                "y": y,
-                "spatialReference": {
-                    "wkid": wkid
+        return __awaiter(this, void 0, void 0, function () {
+            var markerArray, markerLength, _a, Graphic, PopupTemplate, PictureMarkerSymbol, SimpleMarkerSymbol, x, y, content, icon_url, label, wkid, markerSymbol, point, hasPopupDetails, popupTemplate, graphic;
+            return __generator(this, function (_b) {
+                switch (_b.label) {
+                    case 0:
+                        // ?marker=-117;34;4326;My Title;http://www.daisysacres.com/images/daisy_icon.gif;My location&level=10
+                        // ?marker=-117,34,4326,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
+                        // ?marker=-13044705.25,4036227.41,102100,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
+                        // ?marker=-117,34,,My Title,http://www.daisysacres.com/images/daisy_icon.gif,My location&level=10
+                        // ?marker=-117,34,,,,My location&level=10
+                        // ?marker=-117,34&level=10
+                        // ?marker=10406557.402,6590748.134,2526
+                        if (!marker) {
+                            return [2 /*return*/, promiseUtils_1.reject()];
+                        }
+                        markerArray = _splitURLString(marker);
+                        markerLength = markerArray.length;
+                        if (markerLength < 2) {
+                            return [2 /*return*/, promiseUtils_1.reject()];
+                        }
+                        return [4 /*yield*/, Promise.all([
+                                new Promise(function (resolve_1, reject_1) { require(["esri/Graphic"], resolve_1, reject_1); }).then(__importStar),
+                                new Promise(function (resolve_2, reject_2) { require(["esri/PopupTemplate"], resolve_2, reject_2); }).then(__importStar),
+                                new Promise(function (resolve_3, reject_3) { require(["esri/symbols/PictureMarkerSymbol"], resolve_3, reject_3); }).then(__importStar),
+                                new Promise(function (resolve_4, reject_4) { require(["esri/symbols/SimpleMarkerSymbol"], resolve_4, reject_4); }).then(__importStar)
+                            ])];
+                    case 1:
+                        _a = _b.sent(), Graphic = _a[0], PopupTemplate = _a[1], PictureMarkerSymbol = _a[2], SimpleMarkerSymbol = _a[3];
+                        x = parseFloat(markerArray[0]);
+                        y = parseFloat(markerArray[1]);
+                        content = markerArray[3];
+                        icon_url = markerArray[4];
+                        label = markerArray[5];
+                        wkid = markerArray[2] ? parseInt(markerArray[2], 10) : 4326;
+                        markerSymbol = icon_url
+                            ? new PictureMarkerSymbol.default({
+                                url: icon_url,
+                                height: "32px",
+                                width: "32px"
+                            })
+                            : new SimpleMarkerSymbol.default({
+                                outline: {
+                                    width: 1
+                                },
+                                size: 14,
+                                color: [255, 255, 255, 0]
+                            });
+                        point = new Point_1.default({
+                            x: x,
+                            y: y,
+                            spatialReference: {
+                                wkid: wkid
+                            }
+                        });
+                        hasPopupDetails = content || label;
+                        popupTemplate = hasPopupDetails
+                            ? new PopupTemplate.default({
+                                title: content || null,
+                                content: label || null
+                            })
+                            : null;
+                        graphic = new Graphic.default({
+                            geometry: point,
+                            symbol: markerSymbol,
+                            popupTemplate: popupTemplate
+                        });
+                        return [2 /*return*/, graphic];
                 }
             });
-            var hasPopupDetails = content || label;
-            var popupTemplate = hasPopupDetails ?
-                new PopupTemplate({
-                    "title": content || null,
-                    "content": label || null
-                }) : null;
-            var graphic = new Graphic({
-                geometry: point,
-                symbol: markerSymbol,
-                popupTemplate: popupTemplate
-            });
-            return graphic;
         });
     }
     exports.parseMarker = parseMarker;
@@ -206,7 +265,7 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
         }
         var x = parseFloat(positionArray[0]), y = parseFloat(positionArray[1]), z = parseFloat(positionArray[2]);
         var wkid = positionArray.length === 4 ? parseInt(positionArray[3], 10) : 4326;
-        return new Point({
+        return new Point_1.default({
             x: x,
             y: y,
             z: z,
@@ -220,10 +279,12 @@ define(["require", "exports", "esri/Camera", "esri/core/promiseUtils", "esri/cor
             return null;
         }
         var tiltHeadingArray = headingAndTilt.split(",");
-        return tiltHeadingArray.length >= 0 ? {
-            heading: parseFloat(tiltHeadingArray[0]),
-            tilt: parseFloat(tiltHeadingArray[1])
-        } : null;
+        return tiltHeadingArray.length >= 0
+            ? {
+                heading: parseFloat(tiltHeadingArray[0]),
+                tilt: parseFloat(tiltHeadingArray[1])
+            }
+            : null;
     }
     function _getCameraProperties(camera, headingAndTilt) {
         var cameraPosition = _getCameraPosition(camera);
