@@ -1,20 +1,7 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 /// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
 /// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
 import { subclass, declared, property } from 'esri/core/accessorSupport/decorators';
 import Widget from 'esri/widgets/Widget';
-import Accessor from 'esri/core/Accessor';
 import Share from '../components/Share/Share';
 import ShareFeatures from '../components/Share/Share/ShareFeatures';
 import watchUtils = require('esri/core/watchUtils');
@@ -53,7 +40,7 @@ export interface DetailPanelProps extends esri.WidgetProperties {
 }
 
 @subclass('app.DetailPanel')
-class DetailPanel extends declared(Widget, Accessor) {
+class DetailPanel extends declared(Widget) {
 	//--------------------------------------------------------------------------
 	//
 	//  Properties
