@@ -1,15 +1,3 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 /// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
 /// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
 import { subclass, declared, property } from 'esri/core/accessorSupport/decorators';
@@ -23,7 +11,6 @@ import ApplicationBase from 'ApplicationBase/ApplicationBase';
 import { createMapFromItem, createView, getConfigViewProperties } from 'ApplicationBase/support/itemUtils';
 
 import esri = __esri;
-import MaskLayer from '../utilites/MaskLayer';
 
 const CSS = {
 	miniMap: {
@@ -66,7 +53,7 @@ interface MapPanelProps extends esri.WidgetProperties {
 	isMobileView?: boolean;
 }
 @subclass('app.MapPanel')
-class MapPanel extends declared(Widget, Accessor) {
+class MapPanel extends declared(Widget) {
 	//--------------------------------------------------------------------------
 	//
 	//  Properties
