@@ -1,15 +1,3 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 /// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
 /// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
 import { subclass, declared, property } from 'esri/core/accessorSupport/decorators';
@@ -21,7 +9,6 @@ import { tsx } from 'esri/widgets/support/widget';
 import Slider from 'esri/widgets/Slider';
 import esri = __esri;
 import ApplicationBase from 'ApplicationBase/ApplicationBase';
-import * as promiseUtils from 'esri/core/promiseUtils';
 
 const CSS = {
     panel: 'panel',
@@ -39,7 +26,7 @@ interface RefineResultsProps extends esri.WidgetProperties {
 }
 
 @subclass('app.RefineResults')
-class RefineResults extends declared(Widget, Accessor) {
+class RefineResults extends declared(Widget) {
     //--------------------------------------------------------------------------
     //
     //  Properties
@@ -88,10 +75,12 @@ class RefineResults extends declared(Widget, Accessor) {
             max: maxDistance,
             values: [distance],
             precision,
-            rangeLabelsVisible: true,
+            visibleElements: {
+                labels: true,
+                rangeLabels: true
+            },
             labelInputsEnabled,
             rangeLabelInputsEnabled,
-            labelsVisible: true,
             snapOnClickEnabled: true,
             container: node
         });
@@ -103,7 +92,7 @@ class RefineResults extends declared(Widget, Accessor) {
         locale = locale || "en";
         // Append units to range labels
         distanceSlider.labelFormatFunction = (value: number, type: string) => {
-            if (type === "min" || type === "max" || type === "value") {
+            if (type === "min" || type === "max") {
                 return `${value.toLocaleString(locale)} ${convertedUnits}`;
             } else {
                 return value.toLocaleString(locale);
