@@ -1,21 +1,12 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 /// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
 /// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
 var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
     return function (d, b) {
         extendStatics(d, b);
         function __() { this.constructor = d; }
@@ -28,7 +19,43 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/core/tsSupport/decorateHelper", "esri/core/Accessor", "esri/core/Collection", "esri/core/requireUtils", "esri/core/promiseUtils", "esri/core/accessorSupport/decorators", "esri/geometry/Point", "esri/request", "require", "./ShareItem", "./ShareFeatures", "esri/core/watchUtils", "esri/core/Handles"], function (require, exports, __extends, __decorate, Accessor, Collection, requireUtils, promiseUtils, decorators_1, Point, esriRequest, moduleRequire, ShareItem, ShareFeatures, watchUtils, Handles) {
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/core/tsSupport/decorateHelper", "esri/core/Accessor", "esri/core/Collection", "esri/core/accessorSupport/decorators", "esri/geometry/Point", "esri/request", "./ShareItem", "./ShareFeatures", "esri/geometry/projection", "esri/geometry/SpatialReference"], function (require, exports, __extends, __decorate, Accessor, Collection, decorators_1, Point, esriRequest, ShareItem, ShareFeatures, projection, SpatialReference) {
     "use strict";
     //----------------------------------
     //
@@ -42,34 +69,48 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
     //
     //----------------------------------
     var FACEBOOK_ITEM = new ShareItem({
-        id: 'facebook',
-        name: 'Facebook',
-        urlTemplate: 'https://www.facebook.com/sharer/sharer.php?s=100&p[url]={url}'
+        id: "facebook",
+        name: "Facebook",
+        urlTemplate: "https://www.facebook.com/sharer/sharer.php?s=100&u={url}"
     });
     var TWITTER_ITEM = new ShareItem({
-        id: 'twitter',
-        name: 'Twitter',
-        urlTemplate: 'https://twitter.com/intent/tweet?text={summary}&url={url}'
+        id: "twitter",
+        name: "Twitter",
+        urlTemplate: "https://twitter.com/intent/tweet?text={summary}&url={url}"
+    });
+    var LINKEDIN_ITEM = new ShareItem({
+        id: "linkedin",
+        name: "LinkedIn",
+        urlTemplate: "https://www.linkedin.com/shareArticle?url={url}&title={title}&summary={summary}"
+    });
+    var EMAIL_ITEM = new ShareItem({
+        id: "email",
+        name: "E-mail",
+        urlTemplate: "mailto:?subject={title}&body={summary}%20{url}"
     });
     //----------------------------------
     //
     //  Shorten URL API
     //
     //----------------------------------
-    var SHORTEN_API = 'https://arcg.is/prod/shorten';
+    var SHORTEN_API = "https://arcg.is/prod/shorten";
     var ShareViewModel = /** @class */ (function (_super) {
         __extends(ShareViewModel, _super);
         function ShareViewModel() {
+            //----------------------------------
+            //
+            //  Lifecycle
+            //
+            //----------------------------------
             var _this = _super !== null && _super.apply(this, arguments) || this;
-            _this._handles = new Handles();
             //----------------------------------
             //
             //  Private Variables
             //
             //----------------------------------
-            // Promises for widget state
-            _this._shortenPromise = null;
-            _this._projectionPromise = null;
+            // To keep track of widget state
+            _this._shortening = false;
+            _this._projecting = false;
             //----------------------------------
             //
             //  Properties
@@ -89,28 +130,15 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             _this.shareModalOpened = null;
             //----------------------------------
             //
-            // defaultObjectId
-            //
-            //----------------------------------
-            _this.defaultObjectId = null;
-            //----------------------------------
-            //
-            // attachmentIndex
-            //
-            //----------------------------------
-            _this.attachmentIndex = null;
-            //----------------------------------
-            //
-            // isDefault
-            //
-            //----------------------------------
-            _this.isDefault = null;
-            //----------------------------------
-            //
             //  shareItems
             //
             //----------------------------------
-            _this.shareItems = new ShareItemCollection([FACEBOOK_ITEM, TWITTER_ITEM]);
+            _this.shareItems = new ShareItemCollection([
+                FACEBOOK_ITEM,
+                TWITTER_ITEM,
+                LINKEDIN_ITEM,
+                EMAIL_ITEM
+            ]);
             //----------------------------------
             //
             //  shareFeatures
@@ -125,11 +153,6 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             _this.shareUrl = null;
             return _this;
         }
-        //----------------------------------
-        //
-        //  Lifecycle
-        //
-        //----------------------------------
         ShareViewModel.prototype.destroy = function () {
             this.shareItems.removeAll();
             this.view = null;
@@ -141,10 +164,16 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             //
             //----------------------------------
             get: function () {
-                var ready = this.get('view.ready');
+                var ready = this.get("view.ready");
                 return ready
-                    ? this._projectionPromise ? 'projecting' : this._shortenPromise ? 'shortening' : 'ready'
-                    : this.view ? 'loading' : 'disabled';
+                    ? this._projecting
+                        ? "projecting"
+                        : this._shortening
+                            ? "shortening"
+                            : "ready"
+                    : this.view
+                        ? "loading"
+                        : "disabled";
             },
             enumerable: true,
             configurable: true
@@ -156,56 +185,37 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             //
             //----------------------------------
             get: function () {
-                return "<iframe src=\"" + this
-                    .shareUrl + "\" width=\"600\" height=\"450\" frameborder=\"0\" style=\"border:0\" allowfullscreen></iframe>";
+                return "<iframe src=\"" + this.shareUrl + "\" width=\"600\" height=\"450\" frameborder=\"0\" style=\"border:0\" allowfullscreen></iframe>";
             },
             enumerable: true,
             configurable: true
         });
-        ShareViewModel.prototype.initialize = function () {
-            var _this = this;
-            this._handles.add([
-                watchUtils.watch(this, ['defaultObjectId', 'attachmentIndex'], function () {
-                    if (!_this.isDefault) {
-                        _this.generateUrl();
-                    }
-                })
-            ]);
-        };
         //----------------------------------
         //
         //  Public Methods
         //
         //----------------------------------
         ShareViewModel.prototype.generateUrl = function () {
-            var _this = this;
-            if (!this.isDefault) {
-                return this._generateShareUrl().then(function (url) {
-                    var shortenLink = _this.shareFeatures.shortenLink;
-                    if (shortenLink) {
-                        return _this._shorten(url).then(function (shortenedUrl) {
-                            _this._shortenPromise = null;
-                            _this.notifyChange('state');
-                            _this._set('shareUrl', shortenedUrl);
-                            return promiseUtils.resolve(shortenedUrl);
-                        });
+            return __awaiter(this, void 0, void 0, function () {
+                var url, shortenLink, shortenedUrl;
+                return __generator(this, function (_a) {
+                    switch (_a.label) {
+                        case 0: return [4 /*yield*/, this._generateShareUrl()];
+                        case 1:
+                            url = _a.sent();
+                            shortenLink = this.shareFeatures.shortenLink;
+                            if (!shortenLink) return [3 /*break*/, 3];
+                            return [4 /*yield*/, this._shorten(url)];
+                        case 2:
+                            shortenedUrl = _a.sent();
+                            this._set("shareUrl", shortenedUrl);
+                            return [2 /*return*/, shortenedUrl];
+                        case 3:
+                            this._set("shareUrl", url);
+                            return [2 /*return*/, url];
                     }
-                    _this._set('shareUrl', url);
-                    return promiseUtils.resolve(url);
                 });
-            }
-            else {
-                var shortenLink = this.shareFeatures.shortenLink;
-                if (shortenLink) {
-                    var href = window.location.href;
-                    return this._shorten(href).then(function (shortenedUrl) {
-                        _this._shortenPromise = null;
-                        _this.notifyChange('state');
-                        _this._set('shareUrl', shortenedUrl);
-                        return promiseUtils.resolve(shortenedUrl);
-                    });
-                }
-            }
+            });
         };
         //----------------------------------
         //
@@ -213,54 +223,63 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
         //
         //----------------------------------
         ShareViewModel.prototype._generateShareUrl = function () {
-            var _this = this;
-            var href = window.location.href;
-            // If view is not ready
-            if (!this.get('view.ready')) {
-                return promiseUtils.resolve(href);
-            }
-            // Use x/y values and the spatial reference of the view to instantiate a geometry point
-            var _a = this.view.center, x = _a.x, y = _a.y;
-            var spatialReference = this.view.spatialReference;
-            var centerPoint = new Point({
-                x: x,
-                y: y,
-                spatialReference: spatialReference
-            });
-            // Use pointToConvert to project point. Once projected, pass point to generate the share URL parameters
-            return this._processPoint(centerPoint).then(function (point) {
-                _this._projectionPromise = null;
-                _this.notifyChange('state');
-                return _this._generateShareUrlParams(point);
+            return __awaiter(this, void 0, void 0, function () {
+                var href, _a, x, y, spatialReference, centerPoint, point;
+                return __generator(this, function (_b) {
+                    switch (_b.label) {
+                        case 0:
+                            href = window.location.href;
+                            // If view is not ready
+                            if (!this.get("view.ready")) {
+                                return [2 /*return*/, href];
+                            }
+                            _a = this.view.center, x = _a.x, y = _a.y;
+                            spatialReference = this.view.spatialReference;
+                            centerPoint = new Point({
+                                x: x,
+                                y: y,
+                                spatialReference: spatialReference
+                            });
+                            return [4 /*yield*/, this._processPoint(centerPoint)];
+                        case 1:
+                            point = _b.sent();
+                            return [2 /*return*/, this._generateShareUrlParams(point)];
+                    }
+                });
             });
         };
         ShareViewModel.prototype._processPoint = function (point) {
-            var _a = point.spatialReference, isWGS84 = _a.isWGS84, isWebMercator = _a.isWebMercator;
-            // If spatial reference is WGS84 or Web Mercator, use longitude/latitude values to generate the share URL parameters
-            if (isWGS84 || isWebMercator) {
-                return promiseUtils.resolve(point);
-            }
-            this._projectionPromise = requireUtils.when(moduleRequire, [
-                'esri/geometry/projection',
-                'esri/geometry/SpatialReference'
-            ]);
-            this.notifyChange('state');
-            return this._projectionPromise.then(function (_a) {
-                var projection = _a[0], SpatialReference = _a[1];
-                var outputSpatialReference = new SpatialReference({
-                    wkid: 4326
-                });
-                return projection.load().then(function () {
-                    // Check if client side projection is not supported
-                    if (!projection.isSupported()) {
-                        var point_1 = new Point({
-                            x: null,
-                            y: null
-                        });
-                        return promiseUtils.resolve(point_1);
+            return __awaiter(this, void 0, void 0, function () {
+                var _a, isWGS84, isWebMercator, point_1, outputSpatialReference, projectedPoint;
+                return __generator(this, function (_b) {
+                    switch (_b.label) {
+                        case 0:
+                            _a = point.spatialReference, isWGS84 = _a.isWGS84, isWebMercator = _a.isWebMercator;
+                            // If spatial reference is WGS84 or Web Mercator, use longitude/latitude values to generate the share URL parameters
+                            if (isWGS84 || isWebMercator) {
+                                return [2 /*return*/, point];
+                            }
+                            // Check if client side projection is not supported
+                            if (!projection.isSupported()) {
+                                point_1 = new Point({
+                                    x: null,
+                                    y: null
+                                });
+                                return [2 /*return*/, point_1];
+                            }
+                            outputSpatialReference = new SpatialReference({
+                                wkid: 4326
+                            });
+                            this._projecting = true;
+                            this.notifyChange("state");
+                            return [4 /*yield*/, projection.load()];
+                        case 1:
+                            _b.sent();
+                            projectedPoint = projection.project(point, outputSpatialReference);
+                            this._projecting = false;
+                            this.notifyChange("state");
+                            return [2 /*return*/, projectedPoint];
                     }
-                    var projectedPoint = projection.project(point, outputSpatialReference);
-                    return promiseUtils.resolve(projectedPoint);
                 });
             });
         };
@@ -274,15 +293,17 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             var roundedLat = this._roundValue(latitude);
             var zoom = this.view.zoom;
             var roundedZoom = this._roundValue(zoom);
-            var path = href.split('center')[0];
+            var path = href.split("center")[0];
             // If no "?", then append "?". Otherwise, check for "?" and "="
-            var sep = path.indexOf('?') === -1 ? '?' : path.indexOf('?') !== -1 && path.indexOf('=') !== -1 ? '&' : '';
-            var defaultObjectId = this.defaultObjectId !== null ? "&defaultObjectId=" + this.defaultObjectId : '';
-            var attachmentIndex = this.attachmentIndex !== null ? "&attachmentIndex=" + this.attachmentIndex : '';
-            var shareParams = "" + path + sep + "center=" + roundedLon + "," + roundedLat + "&level=" + roundedZoom + defaultObjectId + attachmentIndex;
-            var type = this.get('view.type');
+            var sep = path.indexOf("?") === -1
+                ? "?"
+                : path.indexOf("?") !== -1 && path.indexOf("=") !== -1
+                    ? "&"
+                    : "";
+            var shareParams = "" + path + sep + "center=" + roundedLon + "," + roundedLat + "&level=" + roundedZoom;
+            var type = this.get("view.type");
             // Checks if view.type is 3D, if so add, 3D url parameters
-            if (type === '3d') {
+            if (type === "3d") {
                 var camera = this.view.camera;
                 var heading = camera.heading, fov = camera.fov, tilt = camera.tilt;
                 var roundedHeading = this._roundValue(heading);
@@ -294,24 +315,30 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             return shareParams;
         };
         ShareViewModel.prototype._shorten = function (url) {
-            var requestOptions = {
-                callbackParamName: 'callback',
-                query: {
-                    longUrl: url,
-                    f: 'json'
-                }
-            };
-            this._shortenPromise = esriRequest(SHORTEN_API, requestOptions);
-            this.notifyChange('state');
-            return this._shortenPromise
-                .catch(function (res) {
-                return res;
-            })
-                .then(function (res) {
-                var shortUrl = res.data && res.data.data && res.data.data.url;
-                if (shortUrl) {
-                    return shortUrl;
-                }
+            return __awaiter(this, void 0, void 0, function () {
+                var request, shortUrl;
+                return __generator(this, function (_a) {
+                    switch (_a.label) {
+                        case 0:
+                            this._shortening = true;
+                            this.notifyChange("state");
+                            return [4 /*yield*/, esriRequest(SHORTEN_API, {
+                                    query: {
+                                        longUrl: url,
+                                        f: "json"
+                                    }
+                                })];
+                        case 1:
+                            request = _a.sent();
+                            this._shortening = false;
+                            this.notifyChange("state");
+                            shortUrl = request.data && request.data.data && request.data.data.url;
+                            if (shortUrl) {
+                                return [2 /*return*/, shortUrl];
+                            }
+                            return [2 /*return*/];
+                    }
+                });
             });
         };
         ShareViewModel.prototype._roundValue = function (val) {
@@ -319,13 +346,13 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
         };
         __decorate([
             decorators_1.property({
-                dependsOn: ['view.ready'],
+                dependsOn: ["view.ready"],
                 readOnly: true
             })
         ], ShareViewModel.prototype, "state", null);
         __decorate([
             decorators_1.property({
-                dependsOn: ['shareUrl'],
+                dependsOn: ["shareUrl"],
                 readOnly: true
             })
         ], ShareViewModel.prototype, "embedCode", null);
@@ -335,15 +362,6 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
         __decorate([
             decorators_1.property()
         ], ShareViewModel.prototype, "shareModalOpened", void 0);
-        __decorate([
-            decorators_1.property()
-        ], ShareViewModel.prototype, "defaultObjectId", void 0);
-        __decorate([
-            decorators_1.property()
-        ], ShareViewModel.prototype, "attachmentIndex", void 0);
-        __decorate([
-            decorators_1.property()
-        ], ShareViewModel.prototype, "isDefault", void 0);
         __decorate([
             decorators_1.property({
                 type: ShareItemCollection
@@ -358,7 +376,7 @@ define(["require", "exports", "esri/core/tsSupport/declareExtendsHelper", "esri/
             decorators_1.property({ readOnly: true })
         ], ShareViewModel.prototype, "shareUrl", void 0);
         ShareViewModel = __decorate([
-            decorators_1.subclass('ShareViewModel')
+            decorators_1.subclass("ShareViewModel")
         ], ShareViewModel);
         return ShareViewModel;
     }(decorators_1.declared(Accessor)));
