@@ -1,94 +1,54 @@
-﻿define({
-  root: ({
-    "load": {
-      "label": "טוען"
+define({
+  "load": {
+    "label": "טוען"
+  },
+  "onboarding": {
+    "title": "ברוך הבא!",
+    "content": "חפש כתובת לקבלת מידע נוסף על המיקום וסביבתו.</br>אם אינך יודע מה הכתובת, השתמש באחת משיטות החיפוש הבאות: <ul><li>לחץ על תיבת החיפוש והקלד כתובת או בחר <b>השתמש במיקום הנוכחי </b><img alt src='./assets/search-tip.png'/></li><li>לחץ בתוך המפה <div><img alt src='./assets/map-tip.png'/></div></li></ul></br> התוצאות ייכללו מידע אודות ישויות שיש בהן עניין."
+  },
+  "map": {
+    "label": "מפה",
+    "description": "נווט במפה בעזרת מקשי החצים כדי לזוז ימינה, שמאלה, למעלה או למטה. השתמש במקשי הפלוס והמינוס כדי להגדיל או להקטין את התצוגה. השתמש בתיבת החיפוש כדי לעיין בתוכן המפה",
+    "miniMapDescription": "מפה המציגה את מיקום השכבה שנבחרה"
+  },
+  "count": "סך הכל",
+  "noFeatures": "לא נמצאו ישויות התואמות לקריטריוני החיפוש",
+  "error": "שגיאה",
+  "licenseError": {
+    "message": "החשבון שלך אינו מורשה להשתמש באפליקציות הניתנות להגדרה שאינן ציבוריות. בקש ממנהל המערכת בארגון שלך להקצות לך סוג משתמש שכולל את Essential Apps או רישיון הרחבה ל-Essential Apps",
+    "title": "ללא רישיון"
+  },
+  "tools": {
+    "clearLocation": "נקה מיקום חיפוש",
+    "clear": "נקה",
+    "filter": "פילטר",
+    "close": "סגור",
+    "info": "מידע",
+    "collapse": "קפל הכל",
+    "open": "הרחב הכל",
+    "slider": "מרחק חיפוש",
+    "results": "תוצאות"
+  },
+  "units": {
+    "kilometers": {
+      "singular": "קילומטר",
+      "plural": "קילומטרים",
+      "abbr": "ק\"מ"
     },
-    "onboarding": {
-      "title": "ברוך הבא!",
-      "content": "חפש כתובת לקבלת מידע נוסף על המיקום וסביבתו.</br>אם אינך יודע מה הכתובת, השתמש באחת משיטות החיפוש הבאות: <ul><li>לחץ על תיבת החיפוש והקלד כתובת או בחר <b>השתמש במיקום הנוכחי </b><img alt src=\'./assets/search-tip.png\'/></li><li>לחץ בתוך המפה <div><img alt src=\'./assets/map-tip.png\'/></div></li></ul></br> התוצאות ייכללו מידע אודות ישויות שיש בהן עניין."
+    "feet": {
+      "singular": "רגל",
+      "plural": "רגל",
+      "abbr": "רגל"
     },
-    "map": {
-      "label": "מפה",
-      "description": "נווט במפה בעזרת מקשי החצים כדי לזוז ימינה, שמאלה, למעלה או למטה. השתמש במקשי הפלוס והמינוס כדי להגדיל או להקטין את התצוגה. השתמש בתיבת החיפוש כדי לעיין בתוכן המפה",
-      "miniMapDescription": "מפה המציגה את מיקום השכבה שנבחרה"
+    "miles": {
+      "singular": "מייל",
+      "plural": "מיילים",
+      "abbr": "מייל"
     },
-    "count": "סך הכל",
-    "noFeatures": "לא נמצאו ישויות התואמות לקריטריוני החיפוש",
-    "error": "שגיאה",
-    "licenseError": {
-      "message": "החשבון שלך אינו מורשה להשתמש באפליקציות הניתנות להגדרה שאינן ציבוריות. בקש ממנהל המערכת בארגון שלך להקצות לך סוג משתמש שכולל את Essential Apps או רישיון הרחבה ל-Essential Apps",
-      "title": "ללא רישיון"
-    },
-    "tools": {
-      "clearLocation": "נקה מיקום חיפוש",
-      "clear": "נקה",
-      "filter": "פילטר",
-      "close": "סגור",
-      "info": "מידע",
-      "collapse": "קפל הכל",
-      "open": "הרחב הכל",
-      "slider": "מרחק חיפוש",
-      "results": "תוצאות"
-    },
-    "units": {
-      "kilometers": {
-        singular: "קילומטר",
-        plural: "קילומטרים",
-        abbr: "ק\\"מ"
-      },
-      "feet": {
-        singular: "רגל",
-        plural: "רגל",
-        abbr: "רגל"
-      },
-      "miles": {
-        singular: "מייל",
-        plural: "מיילים",
-        abbr: "מייל"
-      },
-      "meters": {
-        singular: "מטר",
-        plural: "מטרים",
-        abbr: "מ\\'"
-      }
+    "meters": {
+      "singular": "מטר",
+      "plural": "מטרים",
+      "abbr": "מ'"
     }
-  }),
-  "ar": 1,
-  "bs": 1,
-  "ca": 1,
-  "cs": 1,
-  "da": 1,
-  "de": 1,
-  "el": 1,
-  "es": 1,
-  "et": 1,
-  "fi": 1,
-  "fr": 1,
-  "he": 1,
-  "hi": 1,
-  "hr": 1,
-  "hu": 1,
-  "id": 1,
-  "it": 1,
-  "ja": 1,
-  "ko": 1,
-  "lt": 1,
-  "lv": 1,
-  "nl": 1,
-  "nb": 1,
-  "pl": 1,
-  "pt-br": 1,
-  "pt-pt": 1,
-  "ro": 1,
-  "ru": 1,
-  "sl": 1,
-  "sr": 1,
-  "sv": 1,
-  "th": 1,
-  "tr": 1,
-  "uk": 1,
-  "vi": 1,
-  "zh-cn": 1,
-  "zh-hk": 1,
-  "zh-tw": 1
+  }
 });
