@@ -1,15 +1,3 @@
-/*
-  Copyright 2019 Esri
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.​
-*/
 import { ApplicationConfig } from 'ApplicationBase/interfaces';
 import esri = __esri;
 
@@ -48,18 +36,34 @@ export function moveComponent(props: esriMoveWidgetProps) {
 		props.view.ui.move(node, direction);
 	}
 }
-export async function addBasemap(props: esriWidgetProps) {
+async function addBasemap(props: esriWidgetProps) {
 	const { view, config } = props;
 	const BasemapToggle = await import('esri/widgets/BasemapToggle');
 	if (BasemapToggle) {
 		const bmToggle = new BasemapToggle.default({
 			view
 		});
+
 		if (config.altBasemap) {
-			bmToggle.nextBasemap = config.altBasemap;
+			bmToggle.nextBasemap = await _getBasemap(config.altBasemap) as any;
 		}
 		view.ui.add(bmToggle, config.basemapTogglePosition);
 	}
+}
+export async function _getBasemap(id: string) {
+	const Basemap = await import("esri/Basemap");
+	if (!Basemap) { return; }
+
+	let basemap = Basemap.default.fromId(id);
+
+	if (!basemap) {
+		basemap = await new Basemap.default({
+			portalItem: {
+				id
+			}
+		}).loadAll();
+	}
+	return basemap as any;
 }
 export async function addHome(props: esriWidgetProps) {
 	const { view, config } = props;
@@ -74,14 +78,16 @@ export async function addLegend(props: esriWidgetProps) {
 	if (Legend && Expand) {
 		const legend = new Legend.default({
 			view
-		});
+		}) as __esri.Legend;
 		const expand = new Expand.default({
 			view,
 			group: config.legendPosition,
 			mode: 'floating',
 			content: legend
-		});
-
+		}) as __esri.Expand;
+		if (config.legendOpenAtStart) {
+			expand.expand();
+		}
 		view.ui.add(expand, config.legendPosition);
 		const container = expand.container as HTMLElement;
 		container.classList.add('legend-expand');
