@@ -1,8 +1,9 @@
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -86,7 +87,7 @@ define(["require", "exports", "esri/Graphic", "esri/Color", "esri/views/support/
                         distance = geometryEngine_1.default.distance(location, feature.geometry, unit);
                     }
                     if (feature && feature.attributes) {
-                        feature.attributes.lookupDistance = distance !== null ? distance.toLocaleString(locale, { maximumFractionDigits: 2 }) : null;
+                        feature.attributes.lookupDistance = distance !== null ? distance.toLocaleString(locale, { minimumFractionDigits: 2, maximumFractionDigits: 2 }) : null;
                     }
                 });
                 return [2 /*return*/];
